requirements.txt
streamlit==1.32.2
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langchain-groq==0.1.4
faiss-cpu==1.7.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
langchain-cohere==0.1.5
